Fix Subpasses sample on some platforms
When toggling values in the 'Subpasses' sample, it recreates the render targets with the new parameters. The resource caching system was not taking into account that an ImageView that may have been destroyed and recrated with a different image could have the same ImageView handle in some platforms.

I believe the framework has a deeper issue since when a RenderContext gets recreated, it destroys the respective Image and ImageViews but does not remove them from the resource_cache. That may need to be addressed as well although I don't know where's the best place to do that and will leave it out of the scope of this fix.